🚚 api: [EVALUATION] move service to evaluation context
